requirements.txt
asgiref==3.3.1
coverage==5.4
dj-static==0.0.6
Django==3.1.7
django-stdimage==5.1.1
gunicorn==20.0.4
model-mommy==2.0.0
Pillow==8.1.0
progressbar2==3.53.1
psycopg2-binary==2.8.6
python-utils==2.5.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
